// src/types/meter.ts
export interface MeterData {
    volt_l1: number;
    volt_l2: number;
    volt_l3: number;
    curr_l1: number;
    curr_l2: number;
    curr_l3: number;
    act_pwr_l1: number;
    act_pwr_l2: number;
    act_pwr_l3: number;
    apt_pwr_l1: number;
    apt_pwr_l2: number;
    apt_pwr_l3: number;
    rct_pwr_l1: number;
    rct_pwr_l2: number;
    rct_pwr_l3: number;
    pf_l1: number;
    pf_l2: number;
    pf_l3: number;
    pa_l1: number;
    pa_l2: number;
    pa_l3: number;
    avg_volt_ln: number;
    avg_curr_ln: number;
    total_curr_ln: number;
    total_act_pwr: number;
    total_apt_pwr: number;
    total_rct_pwr: number;
    total_pf: number;
    total_pa: number;
    freq: number;
    imp_wh_last_reset: number;
    exp_wh_last_reset: number;
    imp_varh_last_reset: number;
    exp_varh_last_reset: number;
    vah_last_reset: number;
    ah_last_reset: number;
    total_sys_pwr_dmd: number;
    max_sys_pwr_dmd: number;
    imp_act_pwr_dmd: number;
    imp_max_act_pwr_dmd: number;
    exp_act_pwr_dmd: number;
    max_exp_act_pwr_dmd: number;
    total_apt_pwr_dmd: number;
    max_apt_pwr_dmd: number;
    total_curr_dmd: number;
    max_curr_dmd: number;
    total_rct_pwr_dmd: number;
    max_rct_pwr_dmd: number;
    volt_p_seq: number;
    curr_p_seq: number;
    nat_load: number;
    nat_load_l1: number;
    nat_load_l2: number;
    nat_load_l3: number;
    volt_l1_l2: number;
    volt_l2_l3: number;
    volt_l3_l1: number;
    avg_l2l: number;
    curr_neut: number;
    volt_thd_l1: number;
    volt_thd_l2: number;
    volt_thd_l3: number;
    curr_thd_l1: number;
    curr_thd_l2: number;
    curr_thd_l3: number;
    avg_volt_thd: number;
    avg_curr_thd: number;
    curr_dmd_l1: number;
    curr_dmd_l2: number;
    curr_dmd_l3: number;
    max_curr_dmd_l1: number;
    max_curr_dmd_l2: number;
    max_curr_dmd_l3: number;
    volt_l1_l2_thd: number;
    volt_l2_l3_thd: number;
    volt_l3_l1_thd: number;
    avg_thd_ln: number;
    total_kwh: number;
    total_kvarh: number;
    imp_l1_kwh: number;
    imp_l2_kwh: number;
    imp_l3_kwh: number;
    exp_l1_kwh: number;
    exp_l2_kwh: number;
    exp_l3_kwh: number;
    total_l1_kwh: number;
    total_l2_kwh: number;
    total_l3_kwh: number;
    imp_l1_kvarh: number;
    imp_l2_kvarh: number;
    imp_l3_kvarh: number;
    exp_l1_kvarh: number;
    exp_l2_kvarh: number;
    exp_l3_kvarh: number;
    total_l1_kvarh: number;
    total_l2_kvarh: number;
    total_l3_kvarh: number;
    volt_l1_total_harmonic: number;
    volt_l2_total_harmonic: number;
    volt_l3_total_harmonic: number;
    curr_l1_total_harmonic: number;
    curr_l2_total_harmonic: number;
    curr_l3_total_harmonic: number;
    volt_l1_harmonic: Array<number>;
    volt_l2_harmonic: Array<number>;
    volt_l3_harmonic: Array<number>;
    curr_l1_harmonic: Array<number>;
    curr_l2_harmonic: Array<number>;
    curr_l3_harmonic: Array<number>;
}

export interface MeterState {
    data: MeterData;
}